measuring time without overflow

+++ v3/helper.h
@@ -3,7 +3,7 @@
 void enforceMalloc(void* ptr);
 int validateInt(char* str);
 int countLines(char* fileName);
-double get_time_ms();
+long get_time_ms();
 long get_time_ns();
 void activWait(int waitTimeMillis);
 void strupr(char* p);
